Implement loop for portfolio files
Each piece of work has it's own file inside _portfolio/ Jekyll loops over these and inputs them into the index.html page. Layout changes depending on the 'right?' variable in the front matter of the page.

TODO: Implement that piece of works own HTML into it's page inside _portfolio/ to be used as it's page. This will help organise my projects in the future. All it will take to add a new folder is to create that one file. Avoids the need to dig create extra work for myself.